V.0.5.1 - fixed loadTeamsFromDB method. No more double-writing to DB. - added framework for team diplomacy management (WIP)   Commands not yet implemented. - added XP data field for teams

+++ TeamAdvantage/src/com/github/CubieX/TeamAdvantage/CmdExecutors/ListCmd.java
@@ -27,7 +27,7 @@ public class ListCmd implements ISubCmdExecutor
                   countAll++;
                   lineList.add("" + ChatColor.WHITE + countAll + ". " + team.getName() + ChatColor.YELLOW +
                         " - Leiter: " + ChatColor.WHITE + team.getLeader() + ChatColor.YELLOW + " - Geld: " + ChatColor.WHITE +
-                        team.getMoney() + " " + TeamAdvantage.currencyPlural);                      
+                        ((int)team.getMoney()) + " " + TeamAdvantage.currencyPlural);                      
                }
 
                // send list paginated
